v5 REDIR, now perfect (i hope)

+++ libft/strings/ft_putnbr_fd.c
@@ -6,13 +6,13 @@
 /*   By: aurban <aurban@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/23 16:21:13 by aurban            #+#    #+#             */
-/*   Updated: 2024/01/21 20:58:28 by aurban           ###   ########.fr       */
+/*   Updated: 2024/02/03 09:44:44 by aurban           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_putnbr_fd(int n, int fd)
+ssize_t	ft_putnbr_fd(int n, int fd)
 {
 	char	c;
